fix: resolve 6 code review issues (#216-221)
- #216 [P0] Validate procname in callproc() to prevent SQL injection
- #217 [P1] Add OSError to async ping exception handler
- #218 [P1] Fix TIMESTAMP DataSize to 12 (matches actual wire format)
- #219 [P2] Replace assert statements with explicit InterfaceError checks
- #220 [P2] Remove dead code, no-op assignment, document magic numbers
- #221 [P2] Move duplicated helpers to _cursor_common and _connection_common
