+++ pycubrid/aio/cursor.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import logging
+import re
 import time
 from typing import Any, Sequence
 
@@ -11,10 +12,11 @@ from pycubrid._cursor_common import (
     DescriptionItem,
     DML_BATCH_VERBS,
     extract_first_keyword,
+    _raise_batch_error,
 )
 from pycubrid.constants import CUBRIDStatementType
-from pycubrid.exceptions import DatabaseError, InterfaceError, OperationalError, ProgrammingError
-from pycubrid.error_codes import CAS_ERROR_TO_EXCEPTION, _DEFAULT_SQLSTATE
+from pycubrid.exceptions import InterfaceError, OperationalError, ProgrammingError
+
 from pycubrid.protocol import (
     BatchExecutePacket,
     CloseQueryPacket,
@@ -26,29 +28,8 @@ from pycubrid.protocol import (
 
 _LOGGER = logging.getLogger(__name__)
 
-
-def _raise_batch_error(err: dict[str, Any]) -> None:
-    """Raise the appropriate PEP 249 exception for a batch statement failure."""
-    code = err.get("code", -1)
-    message = err.get("message", "batch execute statement failed")
-    exc_name = CAS_ERROR_TO_EXCEPTION.get(code, "DatabaseError")
-    sqlstate = _DEFAULT_SQLSTATE.get(exc_name, "HY000")
-    from pycubrid.exceptions import (
-        DataError,
-        IntegrityError,
-        InternalError,
-    )
-
-    exc_map = {
-        "DataError": DataError,
-        "IntegrityError": IntegrityError,
-        "InternalError": InternalError,
-        "OperationalError": OperationalError,
-        "ProgrammingError": ProgrammingError,
-        "DatabaseError": DatabaseError,
-    }
-    exc_cls = exc_map.get(exc_name, DatabaseError)
-    raise exc_cls(message, code=code, sqlstate=sqlstate)
+# Identifier validation for stored procedure names (prevents SQL injection).
+_IDENTIFIER_RE = re.compile(r"^[a-zA-Z_][a-zA-Z0-9_.]*$")
 
 
 class AsyncCursor(CursorParamsMixin):
@@ -315,6 +296,9 @@ class AsyncCursor(CursorParamsMixin):
         _ = (size, column)
 
     async def callproc(self, procname: str, parameters: Sequence[Any] = ()) -> Sequence[Any]:
+        """Call a stored procedure and return the original parameters."""
+        if not _IDENTIFIER_RE.match(procname):
+            raise ProgrammingError(f"Invalid stored procedure name: {procname!r}")
         placeholders = ", ".join(["?"] * len(parameters))
         if placeholders:
             sql = "CALL %s(%s)" % (procname, placeholders)
